move router pkg
